requirements.txt
fastapi
uvicorn
openai>=1.6.0
requests
python-dotenv
pytest
